requirements.txt
pandas>=2.1
openpyxl>=3.1
pyyaml
pyyaml>=6.0
pydantic>=2.7
regex>=2024.5.15
Unidecode>=1.3
streamlit>=1.39
# opcional (alto desempenho); pode falhar no Windows sem build tools
# pyahocorasick>=2.1
